Updated UI to have a locked coordinate (lat/long) and a value
This will lock in place what axis of a building east / west is being worked on.
NOTE: THIS PRECLUDES DOING ANY SIDES THAT ARE NOT ABSOLUTELY FOLLOWING A DIRECTION (can only scan on a strict N/S or E/W basis)

Added in initial implementation of JSQuandrantHandler that will parse the JS string for information a quadrant's layout

Added quadrant object that will store all that fun information.